reflex_apps/pyproject.toml
[tool.poetry]
name = "reflex-apps"
version = "0.1.0"
description = ""
authors = ["oliveryh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
reflex = "^0.5.7"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlacodegen = {version = "v3.0.0rc5", extras = ["=citext"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
